+++ Scraps.GUI/Bootstrapper.cs
@@ -90,7 +90,6 @@ namespace Scraps.GUI
                 SkipUpdates = args.Any(a => a == "/SkipUpdate"),
                 AutoReconnect = args.Any(a => a == "/AutoReconnect"),
                 SkipAnnouncements = args.Any(a => a == "/SkipAnnouncements"),
-                UseProxies = args.Any(a => a == "/UseProxies"),
             };
 
             return options;