+++ Scraps.GUI/Extensions/RandomExtensions.cs
@@ -16,24 +16,11 @@
 /// along with this program. If not, see <https://www.gnu.org/licenses/>.
 #endregion License
 
-namespace Scraps.GUI.Constants
+namespace Scraps.GUI.Extensions
 {
-    public class Version
+    public static class RandomExtensions
     {
-        public enum ReleaseTypes
-        {
-            Development,
-            PreRelease,
-            Release
-        }
-
-        public static System.Version AsDotNetVersion() => new(Major, Minor, Patch, Hotfix);
-        public static int Major => 4;
-        public static int Minor => 7;
-        public static int Patch => 0;
-        public static int Hotfix => 0;
-        public static ReleaseTypes ReleaseType => ReleaseTypes.Release;
-        public static string SemVer => $"{Major}.{Minor}.{Patch}.{Hotfix}";
-        public static string Full => $"{SemVer}-{ReleaseType}";
+        public static T RandomListItem<T>(this Random rng, List<T> list) => list[rng.Next(list.Count)];
+        public static T RandomArrayItem<T>(this Random rng, T[] list) => list[rng.Next(list.Length)];
     }
 }