+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 4.7.0.0
+
+- Fixed error message relating to the site being down for maintenance
+- Improved exception logging
+- Added experimental proxy support
+  - Using a VPN may be a better solution
+
 ## 4.6.2.0
 
 - Updated validation when saving settings